requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
Django==3.1.7
django-admin==2.0.1
django-appconf==1.0.4
django-compressor==2.4
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-gunicorn==0.1.1
django-htmlmin==0.11.0
django-model-utils==4.1.1
django-money==1.3.1
django-pipeline==2.0.6
django-six==1.0.4
gunicorn==20.1.0
html5lib==1.1
psycopg2-binary==2.8.6
py-moneyed==0.8.0
python-memcached==1.59
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
screen==1.0.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
validate-email==1.3
webencodings==0.5.1
whitenoise==5.2.0
xlwt==1.3.0
